requirements.txt
asgiref==3.4.0; python_version >= "3.6"
django==3.2.4; python_version >= "3.6"
django-environ-2==2.1.0
djangorestframework==3.12.4; python_version >= "3.6"
djangorestframework-simplejwt==4.7.2
djangorestframework-api-key==2.1.0
gunicorn==20.1.0; python_version >= "3.5"
pip==21.1.3
psycopg2-binary==2.9.1; python_full_version >= "3.4.0"
PyJWT==2.1.0
pytz==2021.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
setuptools==57.0.0
sqlparse==0.4.1; python_version >= "3.6"

